Default home for no args
I thought it was a little unintuitive to have to be able to set a name for the home, so i modified the method to set a home with the name "%&%&defaulthome&%&%" if no args are entered. If when /home is entered with no args also, %&%&defaulthome&%&%" will be the set home.